+++ src/advanced/__tests__/advanced.test.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, renderHook, screen, within } from '@testing-library/react';
-import { CartPage } from '../../refactoring/components/shop/CartPage';
-import { AdminPage } from "../../refactoring/components/admin/AdminPage";
+import { CartPage } from '../../refactoring/pages/CartPage';
+import { AdminPage } from "../../refactoring/pages/AdminPage";
 import { CartItem, Coupon, Discount, Product } from '../../types';
 import * as discountUtils from "../../refactoring/utils/discountUtils";
 import * as cartUtils from "../../refactoring/utils/cartUtils";
